requirements.txt
<<<<<<< HEAD
requests==2.31.0
beautifulsoup4==4.12.2

=======
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt --no-deps --universal
anthropic==0.68.1
    # via open-llm-vtuber (pyproject.toml)
azure-cognitiveservices-speech==1.46.0
    # via open-llm-vtuber (pyproject.toml)
chardet==5.2.0
    # via open-llm-vtuber (pyproject.toml)
duckduckgo-mcp-server==0.1.1
    # via open-llm-vtuber (pyproject.toml)
edge-tts==7.2.3
    # via open-llm-vtuber (pyproject.toml)
fastapi==0.118.0
    # via open-llm-vtuber (pyproject.toml)
groq==0.32.0
    # via open-llm-vtuber (pyproject.toml)
httpx==0.28.1
    # via open-llm-vtuber (pyproject.toml)
langdetect==1.0.9
    # via open-llm-vtuber (pyproject.toml)
letta-client==0.1.324
    # via open-llm-vtuber (pyproject.toml)
loguru==0.7.3
    # via open-llm-vtuber (pyproject.toml)
mcp==1.15.0
    # via open-llm-vtuber (pyproject.toml)
numpy==1.26.4
    # via open-llm-vtuber (pyproject.toml)
onnxruntime==1.23.0
    # via open-llm-vtuber (pyproject.toml)
openai==1.109.1
    # via open-llm-vtuber (pyproject.toml)
pre-commit==4.3.0
    # via open-llm-vtuber (pyproject.toml)
pydub==0.25.1
    # via open-llm-vtuber (pyproject.toml)
pysbd==0.3.4
    # via open-llm-vtuber (pyproject.toml)
pyttsx3==2.99
    # via open-llm-vtuber (pyproject.toml)
pyyaml==6.0.3
    # via open-llm-vtuber (pyproject.toml)
requests==2.32.5
    # via open-llm-vtuber (pyproject.toml)
ruamel-yaml==0.18.15
    # via open-llm-vtuber (pyproject.toml)
ruff==0.13.2
    # via open-llm-vtuber (pyproject.toml)
scipy==1.15.3 ; python_full_version < '3.11'
    # via open-llm-vtuber (pyproject.toml)
scipy==1.16.2 ; python_full_version >= '3.11'
    # via open-llm-vtuber (pyproject.toml)
sherpa-onnx==1.12.14
    # via open-llm-vtuber (pyproject.toml)
soundfile==0.13.1
    # via open-llm-vtuber (pyproject.toml)
tomli==2.2.1
    # via open-llm-vtuber (pyproject.toml)
torch==2.2.2 ; platform_machine == 'x86_64' and sys_platform == 'darwin'
    # via open-llm-vtuber (pyproject.toml)
torch==2.8.0 ; platform_machine == 'arm64' or sys_platform != 'darwin'
    # via open-llm-vtuber (pyproject.toml)
tqdm==4.67.1
    # via open-llm-vtuber (pyproject.toml)
uvicorn==0.37.0
    # via open-llm-vtuber (pyproject.toml)
websocket-client==1.8.0
    # via open-llm-vtuber (pyproject.toml)
>>>>>>> main
